Miscellaneous TMC changes

+++ Marlin/src/feature/tmc_util.h
@@ -141,7 +141,7 @@ class TMCMarlin<TMC2208Stepper> : public TMC2208Stepper, public TMCStorage {
     }
 };
 
-constexpr uint32_t _tmc_thrs(const uint16_t msteps, const int32_t thrs, const uint32_t spmm) {
+constexpr uint16_t _tmc_thrs(const uint16_t msteps, const int32_t thrs, const uint32_t spmm) {
   return 12650000UL * msteps / (256 * thrs * spmm);
 }
